+++ ElvUI_MerathilisUI/Core/Changelog/7.08.lua
@@ -1,8 +1,14 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[708] = {
-	RELEASE_DATE = "TBD",
-	FIXES = {},
+	RELEASE_DATE = "31.01.2026",
+	FIXES = {
+		"[Tags]: Use my new/old name tag format [name:MER:gradient] instead of [name:abbrev:medium] in various unitframes.",
+		"        Note: It will be reaction color in instanced content, not a gradient.",
+		"[UnitFrames]: Bring back Power Bar gradient texture",
+		"[Install]: Both install steps from the UnitFrame Layout should work again",
+		"[NamePlates]: CastBar Shield icon was missing, now it's back",
+	},
 	NEW = {},
 	IMPROVEMENTS = {},
 }