+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Changes ###
 
-  * [Portraits]: Fixed some copy-paste errors in the code
-  * [UnitFrames]: Fixed an error with castbars
-  * [UnitFrames]: Remove the Tag [name:gradient] for now and replace it with [name:abbrev:medium]
+  * [Tags]: Use my new/old name tag format [name:MER:gradient] instead of [name:abbrev:medium] in various unitframes
+  *         Note: It will be reaction color in instanced content, not a gradient.
+  * [UnitFrames]: Bring back Power Bar gradient texture
+  * [Install]: Both install steps from the UnitFrame Layout should work again
+  * [NamePlates]: CastBar Shield icon was missing, now it's back